04_done,somehow slow, 9/10

+++ 04_segmBelts/definitions.hpp
@@ -1,4 +1,4 @@
 // normal int
 typedef uint32_t nint;
-typedef uint32_t sint;
+typedef uint16_t sint;
 typedef std::pair<std::string,std::string> pSS;
